chore(arkmkarket): move queries to proper dir
