+++ apps/arkmarket/src/components/token/toast-executed-transaction-content.tsx
@@ -1,4 +1,4 @@
-import type { WalletToken } from "~/lib/getWalletData";
+import type { WalletToken } from "~/queries/getWalletData";
 import type { CollectionToken, Token } from "~/types";
 import Media from "~/components/media";
 import usePrices from "~/hooks/usePrices";
