+++ apps/arkmarket/src/components/token/toast-rejected-transaction-content.tsx
@@ -1,6 +1,6 @@
 import { FileSignature } from "@ark-market/ui/icons";
 
-import type { WalletToken } from "~/lib/getWalletData";
+import type { WalletToken } from "~/queries/getWalletData";
 import type { CollectionToken, Token } from "~/types";
 import Media from "~/components/media";
 import usePrices from "~/hooks/usePrices";
