+++ apps/arkmarket/src/components/wallet/portfolio-items-filters-content.tsx
@@ -16,13 +16,13 @@ import {
 import { VerifiedIcon } from "@ark-market/ui/icons";
 import { Input } from "@ark-market/ui/input";
 
-import type { WalletCollectionsApiResponse } from "~/lib/getWalletData";
+import type { WalletCollectionsApiResponse } from "~/queries/getWalletData";
 import Media from "~/components/media";
-import { getWalletCollections } from "~/lib/getWalletData";
 import {
   walletCollectionFilterKey,
   walletCollectionFilterParser,
 } from "~/lib/wallet-search-params";
+import { getWalletCollections } from "~/queries/getWalletData";
 
 interface PortfolioItemsFiltersContentProps {
   walletAddress: string;
