+++ apps/arkmarket/src/queries/getCollection.ts
