+++ apps/arkmarket/src/queries/getCollectionSearch.ts
