+++ apps/arkmarket/src/queries/getPrices.ts
