+++ apps/arkmarket/src/queries/getSystemStatus.ts
