+++ apps/arkmarket/src/queries/getToken.ts
