+++ apps/arkmarket/src/queries/getTokenMarketData.ts
