+++ apps/arkmarket/src/queries/getWalletData.ts
